+++ app/src/main/java/com/example/project02_triviaapp/database/UserDAO.java
@@ -27,6 +27,6 @@ public interface UserDAO {
     LiveData<User> getUserByUserName(String username);
 
     @Query("SELECT * FROM " + TriviaDatabase.USER_TABLE + " WHERE userid = :userId")
-    LiveData<User> getUserByUserId(int userId);
+    LiveData<User> getUserByUserId(long userId);
 
 }
